Create Exercise 3
This Java program defines a `HelloObject` class, which stores and prints a greeting message provided at initialization. The `HelloTester` class's `main` method prompts the user to enter a greeting, which is then passed to the `HelloObject` constructor. Finally, the `speak` method of `HelloObject` is called to display the entered greeting.