Create Exercise 4
This Java program defines a class `HelloObject` with two constructors and a method to print a message stored in its instance variable `message`. The first constructor initializes the `message` with a provided string value, while the second constructor initializes it with the same value as another `HelloObject` object. The `HelloTester` class tests the functionality by creating instances of `HelloObject`, setting and retrieving messages, demonstrating the use of both constructors.